requirements.txt
Flask
python-dotenv
opencv-python-headless
numpy